+++ src/sort/insertion_sort.rs
@@ -14,12 +14,12 @@ pub fn sort<T: PartialOrd + Copy>(mut array: Vec<T>) -> SortResult<T> {
         .progress_chars("#>-"),
     );
 
-    let mut comparisons: usize = 0;
-    let swaps: usize = 0;
-    let mut shifts: usize = 0;
-    let mut insertions: usize = 0;
-    let moves: usize = 0;
-    let attempts: usize = 0;
+    let mut comparisons: u64 = 0;
+    let swaps: u64 = 0;
+    let mut shifts: u64 = 0;
+    let mut insertions: u64 = 0;
+    let moves: u64 = 0;
+    let attempts: u64 = 0;
 
     let start = Instant::now();
 
@@ -50,7 +50,7 @@ pub fn sort<T: PartialOrd + Copy>(mut array: Vec<T>) -> SortResult<T> {
         pb.set_position(i as u64);
     }
 
-    let duration: usize = start.elapsed().as_nanos() as usize;
+    let duration: u128 = start.elapsed().as_nanos();
     pb.finish_with_message("Sorting completed!");
 
     SortResult {
@@ -72,10 +72,10 @@ mod tests {
 
     #[test]
     fn sort_array() {
-        let array: Vec<i32> = (1..=10).rev().collect();
+        let array: Vec<u64> = (1..=10).rev().collect();
 
         let result = sort(array);
-        let expected: Vec<i32> = (1..=10).collect();
+        let expected: Vec<u64> = (1..=10).collect();
 
         assert_eq!(
             result.array, expected,
